pyproject.toml
[project]
name = "towa-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.69.0",
    "fastapi>=0.118.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.22.0",
    "twelvelabs>=1.0.2",
    "uvicorn>=0.37.0",
    "vapi-server-sdk>=1.7.2",
]
